MAT-9890: clean up compositeMeasureIds on component measures when composite measure is deleted
