+++ src/main/java/cms/gov/madie/measure/services/GroupService.java
@@ -8,7 +8,6 @@ import cms.gov.madie.measure.utils.MeasureUtil;
 import cms.gov.madie.measure.validations.CqlDefinitionReturnTypeService;
 import cms.gov.madie.measure.validations.CqlObservationFunctionService;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import gov.cms.madie.models.common.ActionType;
 import gov.cms.madie.models.common.ModelType;
 import gov.cms.madie.models.measure.Component;
 import gov.cms.madie.models.measure.Group;
@@ -32,12 +31,9 @@ import org.springframework.util.CollectionUtils;
 
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
-import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.stream.Collectors;
 
@@ -55,6 +51,7 @@ public class GroupService {
   private final TestCaseLockService testCaseLockService;
   private final AppConfigService appConfigService;
   private final ActionLogService actionLogService;
+  private final CompositeRelationshipService compositeRelationshipService;
 
   public Group createOrUpdateGroup(Group group, String measureId, String username) {
 
@@ -142,7 +139,7 @@ public class GroupService {
     measureRepository.save(measure);
 
     if (MeasureScoring.COMPOSITE.toString().equalsIgnoreCase(group.getScoring())) {
-      updateComponentCompositeRelationship(
+      compositeRelationshipService.syncComponents(
           previousComponents, group.getComponents(), measure, username);
     }
 
@@ -249,7 +246,8 @@ public class GroupService {
         group -> {
           if (MeasureScoring.COMPOSITE.toString().equalsIgnoreCase(group.getScoring())
               && !CollectionUtils.isEmpty(group.getComponents())) {
-            updateComponentCompositeRelationship(group.getComponents(), List.of(), saved, username);
+            compositeRelationshipService.syncComponents(
+                group.getComponents(), List.of(), saved, username);
           }
         });
 
@@ -499,102 +497,6 @@ public class GroupService {
               .toList();
       testCase.setGroupPopulations(remainingGroups);
     }
-  }
-
-  private void updateComponentCompositeRelationship(
-      List<Component> previousComponents,
-      List<Component> newComponents,
-      Measure compositeMeasure,
-      String username) {
-
-    Set<String> previousIds =
-        previousComponents.stream().map(Component::getMeasureId).collect(Collectors.toSet());
-    Set<String> newIds =
-        newComponents == null
-            ? Set.of()
-            : newComponents.stream().map(Component::getMeasureId).collect(Collectors.toSet());
-
-    Set<String> addedIds =
-        newIds.stream().filter(id -> !previousIds.contains(id)).collect(Collectors.toSet());
-    Set<String> removedIds =
-        previousIds.stream().filter(id -> !newIds.contains(id)).collect(Collectors.toSet());
-
-    Set<String> allChangedIds = new HashSet<>(addedIds);
-    allChangedIds.addAll(removedIds);
-
-    String compositeMeasureId = compositeMeasure.getId();
-    String compositeName = compositeMeasure.getMeasureName();
-
-    // Fetch all added and removed components
-    Map<String, Measure> componentById =
-        measureRepository.findAllById(allChangedIds).stream()
-            .collect(Collectors.toMap(Measure::getId, m -> m));
-
-    // Validate that all component measures exist first
-    allChangedIds.forEach(
-        id -> {
-          if (!componentById.containsKey(id)) {
-            throw new ResourceNotFoundException("Measure", id);
-          }
-        });
-
-    // Update all components in memory before writing to db
-    addedIds.forEach(
-        id -> {
-          Measure component = componentById.get(id);
-          List<String> ids =
-              new ArrayList<>(
-                  Objects.requireNonNullElseGet(
-                      component.getCompositeMeasureIds(), ArrayList::new));
-          ids.add(compositeMeasureId);
-          component.setCompositeMeasureIds(ids);
-        });
-
-    removedIds.forEach(
-        id -> {
-          Measure component = componentById.get(id);
-          List<String> ids =
-              new ArrayList<>(
-                  Objects.requireNonNullElseGet(
-                      component.getCompositeMeasureIds(), ArrayList::new));
-          ids.remove(compositeMeasureId);
-          component.setCompositeMeasureIds(ids);
-        });
-
-    measureRepository.saveAll(componentById.values());
-
-    // Log after all writes succeed
-    addedIds.forEach(
-        id -> {
-          actionLogService.logAction(
-              compositeMeasureId,
-              Measure.class,
-              ActionType.COMPONENT_ADDED,
-              username,
-              "Added Component measure " + componentById.get(id).getMeasureName());
-          actionLogService.logAction(
-              id,
-              Measure.class,
-              ActionType.ADDED_TO_COMPOSITE,
-              username,
-              "Added to Composite measure " + compositeName);
-        });
-
-    removedIds.forEach(
-        id -> {
-          actionLogService.logAction(
-              compositeMeasureId,
-              Measure.class,
-              ActionType.COMPONENT_REMOVED,
-              username,
-              "Removed Component measure " + componentById.get(id).getMeasureName());
-          actionLogService.logAction(
-              id,
-              Measure.class,
-              ActionType.REMOVED_FROM_COMPOSITE,
-              username,
-              "Removed from Composite measure " + compositeName);
-        });
   }
 
   protected void handleFhirGroupReturnTypes(Group group, Measure measure) {
